requirements.txt
flask
gunicorn
pillow
requests
python-dotenv
